heroku deploy

+++ backend/src/routers/cityUserInputRouter.ts
@@ -93,6 +93,22 @@ cityUserInputRouter.put("/:id", verifyToken, async (req, res) => {
     anonymous,
   } = req.body;
 
+  const item = await prisma.cityUserInput.findUnique({
+    where: { id: Number(req.params.id) }, // Assuming `id` is the primary key
+    select: { userId: true }, // Adjust based on your model's fields
+  });
+
+  if (!item) {
+    return res.status(404).json({ error: `suggestion not found` });
+  }
+
+  // Check if the current user is the owner of the item
+  if (item.userId !== req.user.id) {
+    return res
+      .status(403)
+      .json({ error: "You do not have permission to update this item" });
+  }
+
   try {
     const updatedCityUserInput = await prisma.cityUserInput.update({
       where: { id: Number(id) },
@@ -123,6 +139,22 @@ cityUserInputRouter.put("/:id", verifyToken, async (req, res) => {
 // Delete a CityUserInput by ID
 cityUserInputRouter.delete("/:id", verifyToken, async (req, res) => {
   const { id } = req.params;
+
+  const item = await prisma.cityUserInput.findUnique({
+    where: { id: Number(req.params.id) }, // Assuming `id` is the primary key
+    select: { userId: true }, // Adjust based on your model's fields
+  });
+
+  if (!item) {
+    return res.status(404).json({ error: `city user input not found` });
+  }
+
+  // Check if the current user is the owner of the item
+  if (item.userId !== req.user.id) {
+    return res
+      .status(403)
+      .json({ error: "You do not have permission to update this item" });
+  }
 
   try {
     await prisma.cityUserInput.delete({
